Moves toJson to DelayStats.
It uses fields that are internal and the usage in java is clumsy:
JsonStats.toJson(DtlsTransport.packetDelayStats)
