--- src/main/kotlin/org/jitsi/nlj/stats/JsonStats.kt
@@ -1,36 +1,0 @@
-/*
- * Copyright @ 2018 - present 8x8, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-@file:JvmName("JsonStats")
-
-package org.jitsi.nlj.stats
-
-import org.jitsi.nlj.util.OrderedJsonObject
-
-fun DelayStats.toJson(): OrderedJsonObject {
-    return OrderedJsonObject().apply {
-        put("average_delay_ms", averageDelay)
-        put("max_delay_ms", maxDelayMs)
-
-        val buckets = OrderedJsonObject().apply {
-            for (i in thresholds.indices) {
-                put("< ${thresholds[i]} ms", thresholdCounts[i])
-            }
-            put("> ${thresholds[thresholds.size - 1]} ms", thresholdCounts[thresholds.size])
-        }
-        put("buckets", buckets)
-    }
-}
